// src/utils/api/error/apiErrorCode.ts
export enum ApiErrorCode {
  INTERNAL = 0,
  ILLEGAL_ARGUMENT = 2,
  TOKEN_NOT_FOUND = 11,
  TOKEN_EXPIRED = 12,
  STORAGE_EXCEPTION = 19,
  STORAGE_FILE_NOT_FOUND = 20,
  IMAGE_ALREADY_LINKED = 21,
  EMAIL_IS_ALREADY_USED = 22,
  SHOP_NAME_IS_ALREADY_USED = 23,
  CLIENT_ALREADY_CREATED = 24,
  USER_NOT_FOUND = 34,
  CLIENT_NOT_FOUND = 35,
  ACCESS_DENIED = 37,
  UNACCEPTABLE_EMAIL = 41,
  NO_SHOP_REGISTERED = 50,
}